requirements.txt
httpx 
python-dotenv
fastapi
uvicorn
sqlalchemy 
alembic 
psycopg2-binary
torch
